Squashed commit of the following:
commit 0a64395544c18be9ba7dc92101ae5c3ade8ccced
Author: leoafarias <leokcta@gmail.com>
Date:   Thu Sep 17 12:40:07 2020 -0400

    Improvements

commit deea87182e97189ba7b904510f1a3a7e72f65fc6
Author: leoafarias <leokcta@gmail.com>
Date:   Thu Sep 17 12:35:54 2020 -0400

    Tweaks

commit 0e9aca391ae10f73aa5b93e29cdf7a7f8223e587
Author: leoafarias <leokcta@gmail.com>
Date:   Thu Sep 17 12:33:11 2020 -0400

    Improvements

commit 57b354a09bc9059a917dee827d94290cc9449eb7
Author: leoafarias <leokcta@gmail.com>
Date:   Thu Sep 17 12:24:52 2020 -0400

    Test

commit c0140c5b719a9e9122fd046b98c6845a9e777f99
Author: leoafarias <leokcta@gmail.com>
Date:   Thu Sep 17 12:21:19 2020 -0400

    Adjustment

commit 41feef73b897973ed521a860be372eb0dd800b64
Author: leoafarias <leokcta@gmail.com>
Date:   Thu Sep 17 12:16:59 2020 -0400

    Test adjustments

commit ef5b599dbd72fe5a63314de4acf07254530f5a6d
Author: leoafarias <leokcta@gmail.com>
Date:   Thu Sep 17 08:35:53 2020 -0400

    CI adjustments

commit 03d796c67c43642e3a15232a5703e6383168e520
Author: leoafarias <leokcta@gmail.com>
Date:   Thu Sep 17 06:06:42 2020 -0400

    Removed verbose

commit 41e84509904f39aa87a0231f3bd9c7440e9648bb
Author: leoafarias <leokcta@gmail.com>
Date:   Thu Sep 17 06:04:58 2020 -0400

    added verbose

commit 084e42c7501ac04505e8939d274198975515c319
Author: leoafarias <leokcta@gmail.com>
Date:   Thu Sep 17 06:03:36 2020 -0400

    Test changes

commit 578cdd82d8dedd301ada241f566e68beea1a3455
Author: leoafarias <leokcta@gmail.com>
Date:   Thu Sep 17 05:52:05 2020 -0400

    Removed fail-fast

commit 63abff4441bd704c5ab4182f82526409de4419ff
Author: leoafarias <leokcta@gmail.com>
Date:   Thu Sep 17 05:50:20 2020 -0400

    Removed flutter setup

commit c330b6ed3b2ecb435e959fc04250c385dc0fee81
Author: leoafarias <leokcta@gmail.com>
Date:   Thu Sep 17 05:48:30 2020 -0400

    Further tweaks

commit 541e4ac95cfe33e5707f313c36bebb0b2bfd3a63
Author: leoafarias <leokcta@gmail.com>
Date:   Thu Sep 17 05:39:23 2020 -0400

    Added multiple OSs

+++ packages/cli/lib/src/flutter_tools/git_tools.dart
@@ -16,7 +16,6 @@ Future<void> _checkIfGitInstalled() async {
       'git',
       ['--version'],
       workingDirectory: kWorkingDirectory.path,
-      runInShell: Platform.isWindows,
     );
   } on ProcessException {
     throw Exception(
@@ -27,6 +26,7 @@ Future<void> _checkIfGitInstalled() async {
 /// Clones Flutter SDK from Version Number or Channel
 /// Returns exists:true if comes from cache or false if its new fetch.
 Future<void> runGitClone(String version) async {
+  print('RUNNING GIT CLONE');
   await _checkIfGitInstalled();
   final versionDirectory = Directory(path.join(kVersionsDir.path, version));
 
@@ -47,32 +47,38 @@ Future<void> runGitClone(String version) async {
   final process = await run(
     'git',
     args,
-    stdout: consoleController.stdoutSink,
-    stderr: consoleController.stderrSink,
-    runInShell: Platform.isWindows,
+    commandVerbose: true,
+    stdout: consoleController.stdout,
+    stderr: consoleController.stderr,
     verbose: logger.isVerbose,
+    runInShell: Platform.isWindows,
   );
 
   if (process.exitCode != 0) {
-    // If clone to directory fails. Remove directory
-    if (await versionDirectory.exists()) {
-      await versionDirectory.delete();
+    // Did not cleanly exit clean up directory
+    if (process.exitCode == 128) {
+      if (await versionDirectory.exists()) {
+        await versionDirectory.delete();
+      }
     }
     throw InternalError('Could not install Flutter version: $version.');
   }
+
+  return;
 }
 
 Future<String> getCurrentGitBranch(Directory dir) async {
   try {
     if (!await dir.exists()) {
-      throw Exception('Could not get version from SDK that is not installed');
+      throw Exception(
+          'Could not get GIT version from ${dir.path} that does not exist');
     }
     var result = await run('git', ['rev-parse', '--abbrev-ref', 'HEAD'],
-        workingDirectory: dir.path, runInShell: true);
+        workingDirectory: dir.path);
 
     if (result.stdout.trim() == 'HEAD') {
       result = await run('git', ['tag', '--points-at', 'HEAD'],
-          workingDirectory: dir.path, runInShell: true);
+          workingDirectory: dir.path);
     }
 
     if (result.exitCode != 0) {
@@ -83,6 +89,14 @@ Future<String> getCurrentGitBranch(Directory dir) async {
   } on Exception catch (err) {
     //TODO: better error logging
     FvmLogger.error(err.toString());
+    return null;
+  }
+}
+
+Future<String> getProjectGitBranch(Directory dir) async {
+  try {
+    return await getCurrentGitBranch(dir);
+  } on Exception {
     return null;
   }
 }
